+++ lib/src/cmsd/CMSD_DYNAMIC.ts
@@ -1,0 +1,8 @@
+/**
+ * CMSD dynamic header name.
+ * 
+ * @group CMSD
+ * 
+ * @beta
+ */
+export const CMSD_DYNAMIC = 'CMSD-Dynamic';