+++ lib/src/index.ts
@@ -4,5 +4,9 @@
  * @packageDocumentation
  */
 export * from './cmcd.js';
+export * from './cmsd.js';
+export type * from './cta.js';
 export * from './id3.js';
+export type * from './request.js';
+export * from './structuredfield.js';
 export * from './utils.js';