+++ lib/src/structuredfield/parse/ParsedValue.ts
@@ -1,0 +1,4 @@
+export type ParsedValue<T> = {
+  value: T;
+  src: string;
+}