+++ lib/src/utils.ts
@@ -3,5 +3,8 @@
  * 
  * @packageDocumentation
  */
+export * from './utils/base64decode.js';
+export * from './utils/base64encode.js';
+export * from './utils/roundToEven.js';
 export * from './utils/utf8ArrayToStr.js';
 export * from './utils/uuid.js';