+++ lib/test/cmsd/data/CMSD_DYNAMIC_SINGLE.ts
@@ -1,0 +1,1 @@
+export const CMSD_DYNAMIC_SINGLE = `"CDNB-3ak1";etp=96;rtt=8`;