+++ CHANGELOG.md
@@ -19,6 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.5.0] - 2023-??-??
+
+### Added
+- Implement Common Media Server Data (CMSD) utilities [#35](https://github.com/streaming-video-technology-alliance/common-media-library/issues/35)
+- Add request and response interceptors API
+
+
 ## [0.4.5] - 2023-10-12
 
 ### Fixed
@@ -133,7 +140,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bootstrap project [#2](https://github.com/streaming-video-technology-alliance/common-media-library/issues/2)
 
 
-[Unreleased]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.4.5...HEAD  
+[Unreleased]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.5.0...HEAD  
+[0.5.0]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.4.5...v0.5.0  
 [0.4.5]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.4.4...v0.4.5  
 [0.4.4]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.4.3...v0.4.4  
 [0.4.3]\: https://github.com/streaming-video-technology-alliance/common-media-library/compare/v0.4.2...v0.4.3